stm32/(q|o|x)spi: Change the default device size to maximum allowable.
It is relatively easy to get caught by the fact that the default memory
size will limit the indirect read size to maximum of 2 bytes. The memory
size has to be set to something bigger for the indirect read to allow
for more bytes being read out of the device.