nrf/i2s: allow master mode without MCK pin
Outputting the generated master clock is optional as per Nordic docs:
"The MCK signal can be routed to an output pin (specified in PSEL.MCK)
to supply external I2S devices that require the MCK to be supplied from
the outside." [1]

Some I2S DACs, such as MAX98357A, don't require a master clock input so
it can be left disconnected. nRF Connect SDK supports this.

[1] https://docs.nordicsemi.com/bundle/ps_nrf52833/page/i2s.html
[2] https://www.analog.com/media/en/technical-documentation/data-sheets/max98357a-max98357b.pdf